msm: camera: isp: Set vfe clock rate
Set the initial vfe clock rate to SVS during probe. When the
clock is enabled the clock driver expects a clock rate to be
voted. The clock is enabled during device open and the
user driver can send the actual rate only after open. This causes
a warning print from clock driver when the clock is enabled in
device open. To avoid the error print set an initail rate.

CRs-Fixed: 1000316
Change-Id: I423469eef79fc9251bab52ea3a3b3c9d8db422a3
Signed-off-by: Shubhraprakash Das <sadas@codeaurora.org>
Signed-off-by: Lau <laststandrighthere@gmail.com>
Signed-off-by: GhostMaster69-dev <rathore6375@gmail.com>
Signed-off-by: rk134 <rk134@pixelexperience.org>